+++ src/modules/purchasing/purchase-request/view.js
@@ -1,9 +1,9 @@
 import {inject, Lazy} from 'aurelia-framework';
 import {Router} from 'aurelia-router';
 import {Service} from './service';
+import { Dialog } from '../../../au-components/dialog/dialog';
 
-
-@inject(Router, Service)
+@inject(Router, Service, Dialog)
 export class View {
     hasCancel = true;
     hasEdit = false;
@@ -11,9 +11,10 @@ export class View {
     hasUnpost = false;
     prId = "";
 
-    constructor(router, service) {
+    constructor(router, service, dialog) {
         this.router = router;
         this.service = service;
+        this.dialog = dialog;
     }
 
     async activate(params) {
@@ -71,9 +72,15 @@ export class View {
     }
 
     delete(event) {
-        this.service.delete(this.data).then(result => {
-            this.cancel();
-        });
+        this.dialog.prompt('Apakah anda yakin akan menghapus data ini?', 'Hapus Data Purchase Request')
+        .then(response => {
+            if (response.ok) {
+                this.service.delete(this.data).then(result => {
+                    alert("Data berhasil dihapus");
+                    this.cancel();
+                });
+            }
+        })
     }
 
     unpost(event) {
