+++ src/modules/purchasing/unit-receipt-note/view.js
@@ -1,13 +1,14 @@
 import {inject, Lazy} from 'aurelia-framework';
 import {Router} from 'aurelia-router';
 import {Service} from './service';
+import { Dialog } from '../../../au-components/dialog/dialog';
 
-
-@inject(Router, Service)
+@inject(Router, Service, Dialog)
 export class View {
-    constructor(router, service) {
+    constructor(router, service, dialog) {
         this.router = router;
         this.service = service;
+        this.dialog = dialog;
     }
 
     async activate(params) {
@@ -48,15 +49,24 @@ export class View {
     }
 
     delete() {
-        this.service.delete(this.data).then(result => {
-          if (result != "") {
-            alert(result)
-          }
-          else {
-            alert("Data berhasil dihapus");
-          }
-          this.list();
-        });
+        // this.service.delete(this.data).then(result => {
+        //   if (result != "") {
+        //     alert(result)
+        //   }
+        //   else {
+        //     alert("Data berhasil dihapus");
+        //   }
+        //   this.list();
+        // });
+        this.dialog.prompt('Apakah anda yakin akan menghapus data ini?', 'Hapus Data Bon Terima Unit')
+        .then(response => {
+            if (response.ok) {
+                this.service.delete(this.data).then(result => {
+                    alert("Data berhasil dihapus");
+                    this.list();
+                });
+            }
+        })
     }
 
     showDetail(item) {
